MathInputButtons with customized placeholders: what you see is what you get in the output
In case MathInputButtons have the same TextColor as the output, the placeholders can be customized by only setting one or more of the LaTeXSetting properties PlaceholderActiveNucleus, PlaceholderRestingNucleus, PlaceholderActiveColor, PlaceholderRestingColor.
In case of different keyboard colors than output colors then the MathInputButton properties PlaceholderActiveColor and/or PlaceholderRestingColor can be used to override the LaTeXSettings. Those two properties use a weird hack: BindableProperties cannot have nullable value types it seems. Nobody will ever use a transparent placeholder, so that is save to use as a replacement for null. The color black should be available, in case the TextColor is not black.

This commit reverts the adding of MyMathInputButton of commit 94ebda211b7905c4651d2ede4f0b2b3eab9e61fd.

The Example project's Editor tab is updated to have a "Change appearance" button on the EditorPage that does a round trip through 3 themes that show different Clear buttons, different colors on the keyboard keys and output, different placeholders on the keyboard and in the output.

A large number of testcases has been added. It does not cover everything however.
The PlaceholderColorsProperties_MathInputButton expects a weird LaTeX string (you see "{}" at index 24):
@"\(\color{blue}{\square }{}^{\color{green}{■}}\)"